requirements.txt
# AI Team Platform - Core Dependencies
# Install with: pip install ai-team-platform
# Or for all features: pip install ai-team-platform[all]

# Core Framework
mcp>=1.0.0                     # MCP server for Claude Code integration
aiohttp>=3.9.0
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.10.0
jinja2>=3.1.2
python-dotenv>=1.0.0
python-multipart>=0.0.9
requests>=2.31.0
aiofiles>=23.2.1

# AI Model SDKs (install as needed)
anthropic>=0.40.0          # Claude - pip install ai-team-platform[claude]
openai>=1.50.0             # ChatGPT/Grok - pip install ai-team-platform[openai]
google-genai>=1.0.0        # Gemini - pip install ai-team-platform[gemini]

# Integrations (optional)
firebase-admin>=6.0.0      # Firestore - pip install ai-team-platform[firestore]
PyGithub>=2.0.0            # GitHub - pip install ai-team-platform[github]
flask>=3.0.0               # Flask API - pip install ai-team-platform[flask]
flask-cors>=4.0.0

# ML Intelligence (optional) - pip install ai-team-platform[ml]
sentence-transformers>=2.2.0  # Code embeddings
numpy>=1.24.0                  # Array operations
chromadb>=0.4.0                # Local vector DB for RAG
PyYAML>=6.0                    # Project manifest parsing

# Apple Silicon MLX (Mac only) - pip install ai-team-platform[mlx]
mlx>=0.4.0                     # Apple ML framework
mlx-lm>=0.4.0                  # MLX language models
# mlx-embeddings>=0.1.0        # MLX embeddings (when available)

# Cross-platform ML (optional)
# torch>=2.0.0                 # PyTorch (use MLX on Mac instead)

# Development & Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
black>=24.0.0
ruff>=0.2.0
